Update wvc-700-tiny85.ino
The most important change is that the MPP search no longer calculates with the measured current but with the reference value that the controller sets for the DC/DC converter. 

As long as the input voltage is large enough and can be fed in sine-wave, this is directly proportional to the current. In opposite to the current, however, it is not affected by any interferences.

Further, the scaling and the offset of the measured values have been dropped. On closer inspection, this turned out to be mathematically useless and only costs processing time.